+++ js/custom/units.js
@@ -11,28 +11,40 @@ $(document).ready(function() {
         
         
         if(name == "") {
-          $('#reqU').html("<div class='alert alert-danger'>All fields are required</div>");
-          return;
+            $('#reqU').html("<div class='alert alert-danger'>All fields are required</div>");
+            return;
         }
         
-        $.ajax({
-          type: "POST",
-          contentType: "application/json; charset=utf-8",
-          url: api+"create_units.php",
-          data: JSON.stringify({'name': name}),
-          cache: false,
-          success: function(result) {
-            $('#mgs_unts').html('<div class="alert alert-success">Unit added successfully</div>');
-              setTimeout(function () {  
-                   location.reload(true);
-                 }, 1000);
-                  
-          },
-          error: function(err) {
-            alert(err);
-          }
+        // Check if the unit already exists
+        $.getJSON(api + 'read_units.php', function(json) {
+            var unitExists = json.some(function(unit) {
+                return unit.name === name;
+            });
+    
+            if (unitExists) {
+                $('#reqU').html("<div class='alert alert-danger'>Variant already exists in the system</div>");
+                return;
+            }
+    
+            // If the unit doesn't exist, proceed with adding it
+            $.ajax({
+                type: "POST",
+                contentType: "application/json; charset=utf-8",
+                url: api + "create_units.php",
+                data: JSON.stringify({'name': name}),
+                cache: false,
+                success: function(result) {
+                    $('#mgs_unts').html('<div class="alert alert-success">Variant added successfully</div>');
+                    setTimeout(function () {  
+                        location.reload(true);
+                    }, 1000);
+                },
+                error: function(err) {
+                    alert(err);
+                }
+            });
         });
-      });
+    });
 
 
                 ////////////////////////////////////end add/////////////////////////////////
@@ -52,7 +64,7 @@ $(document).ready(function() {
                     var btnEdit = '<button class="btn btn-primary btn-sm btn-rounded btn_edit" data-toggle="modal" data-target="#edit-unit" data-edit="' + json[i].units_id + '"><svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="#fff"><path d="M3 17.25v3.75h3.75L17.81 9.19l-3.75-3.75L3 17.25zm18.71-5.41c.39-.39.39-1.02 0-1.41l-2.34-2.34c-.39-.39-1.02-.39-1.41 0l-1.83 1.83 3.75 3.75 1.83-1.83z"/><path d="M0 0h24v24H0z" fill="none"/></svg>Edit</button>';
                     var btnDelete = '<button class=\'btn btn-primary btn-sm btn-rounded btn-delete btn-del\' data-toggle=\'modal\' data-target=\'#delete-unit\' data-del=' + json[i].units_id  + '><svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-trash" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><path stroke="none" d="M0 0h24v24H0z" fill="none"/><line x1="4" y1="7" x2="20" y2="7" /><line x1="10" y1="11" x2="10" y2="17" /><line x1="14" y1="11" x2="14" y2="17" /><path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12" /><path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3" /></svg> Delete</button>';
 
-
+                    obj['units_id'] = json[i].units_id;
                     obj['name'] = json[i].name;
                     obj['actions'] = btnEdit + " " + btnDelete;
 
@@ -63,6 +75,8 @@ $(document).ready(function() {
 
                         'data': arr,
                         'columns': [{
+                            "data": "units_id"
+                        },{
                             "data": "name"
                         },{
                             "data": "actions"
@@ -113,7 +127,7 @@ $(document).ready(function() {
                      var units_id = $('#edit_unitsid').val();
    
                     if(name == "") {
-                      $('#reqU1').html("<div class='alert alert-danger'>Unit Name is required</div>");
+                      $('#reqU1').html("<div class='alert alert-danger'>Variant Name is required</div>");
                       return;
                     }
                     
@@ -124,7 +138,7 @@ $(document).ready(function() {
                       data: JSON.stringify({'units_id': units_id, 'name': name}),
                       cache: false,
                       success: function(result) {
-                        $('#mgs_uedit').html('<div class="alert alert-success">Unit Updated successfully</div>');
+                        $('#mgs_uedit').html('<div class="alert alert-success">Variant Updated successfully</div>');
                           setTimeout(function () {  
                                location.reload(true);
                              }, 1000);
@@ -167,29 +181,38 @@ $(document).ready(function() {
 
              ////////////////////////////////// delete process////////////////////////////////////////
 
-               $(document).delegate('#del_units', 'click', function(event) {
-                    event.preventDefault();
+             //////////////////////////////// delete process////////////////////////////////////////
 
-                     var units_id = $('#del_unitsid').val();
+            $(document).delegate('#del_units', 'click', function(event) {
+              event.preventDefault();
 
-                    $.ajax({
-                      type: "DELETE",
-                      contentType: "application/json; charset=utf-8",
-                      url: api+"delete_units.php?units_id="+units_id,
-                      cache: false,
-                      success: function(result) {
-                        $('#mgs_Udel').html('<div class="alert alert-success">Unit Deleted successfully</div>');
-                          setTimeout(function () {  
-                               location.reload(true);
-                             }, 1000);
-                              
-                      },
-                      error: function(err) {
-                        alert(err);
-                      }
-                    });
-                  });
-              
+              // Show confirmation dialog
+              var confirmDelete = confirm("Are you sure you want to delete this variant?");
+              if (!confirmDelete) {
+                  return; // Cancel deletion if user chooses No
+              }
+
+              var units_id = $('#del_unitsid').val();
+
+              $.ajax({
+                  type: "DELETE",
+                  contentType: "application/json; charset=utf-8",
+                  url: api + "delete_units.php?units_id=" + units_id,
+                  cache: false,
+                  success: function(result) {
+                      $('#mgs_Udel').html('<div class="alert alert-success">Variant Deleted successfully</div>');
+                      setTimeout(function () {  
+                          location.reload(true);
+                      }, 1000);
+                  },
+                  error: function(err) {
+                      alert(err);
+                  }
+              });
+            });
+
+//////////////////////////////////////////end delete process///////////////////////////////
+
              
               //////////////////////////////////////////end delete process///////////////////////////////
